Add files via upload
Added the notification class, and admin class with the required methods. Also added the fetch_available_appointments functions to Fetches available appointments from an external API and displays them, and adjust_clinic_capacity function to Adjusts the capacity of a clinic by sending a POST request to an external API. These last two functions are not completed and need to be modified accordingly. Also defined 2 functions as test cases to test if what we have done is ok or not for both senarios.